requirements.txt
PyQt6>=6.6.0
pyautogui>=0.9.54
pyperclip>=1.8.2
pytesseract>=0.3.10
Pillow>=10.0.0
openpyxl>=3.1.2
xlrd>=2.0.1
pandas>=2.0.0
selenium>=4.15.0
webdriver-manager>=4.0.1
requests>=2.31.0
pymysql>=1.1.0
psycopg2-binary>=2.9.9
pywin32>=306
pygetwindow>=0.0.9
pyautogui>=0.9.54
mss>=9.0.1
numpy>=1.24.0
APScheduler>=3.10.4
schedule>=1.2.0
python-dotenv>=1.0.0
psutil>=5.9.6
